Refactor assertions
Assertions currently have similar code inside to build failure
messages and generate errors. This is not a good solution in
case something has to change.

A better solution is to use some kind of abstract OOP solution
which allows child classes to change only parts of the process
if necessary.

This commit borrows some aspects from PHPUnit, which has quite
a solid interface thas is used by many projects with great
success. It is also easier to adapt to LV than many other unit
test solutions (I tried a few, but this seems the easiest).

This commit replaces the previous implementation of AssertTrue
and AssertFalse by classes IsTrue and IsFalse, which are child
classes of Constraint (all inside the Constraint library).

The existing tests already cover every functionality that must
be tested (public API).