Implement early test runner
This is an early implementation of a test runner that executes
all tests in the current project by searching for libraries in
the current project that end with 'Test'. All public VIs within
these libraries are assumed to be tests, all of which must have
an error out terminal.

This certainly requires a lot of cleaning up, but it works for
now.