Remove 'Result' terminals
There is a 'Result' output terminal on every assert function
which has the same purpose as the error out terminal. 'Result'
will be false if there is an error, and true if there is none.
It doesn't matter if the error ocurred before the assertion
failed, or during.

This also reflects in the tests, which need to check for both
'Result' and 'error out' terminals to return matching values.

The 'Result' terminal should be removed to avoid complex test
code.

Note that it would also be possible to remove the 'error out'
terminal, but then we don't have a good way to provide failure
messages.

In the LV realm it is best to use what is provided by LV, which
is the error cluster, and simply utilize it for our purposes.