Add test type 'Theory'
A theory is a test case that is presented with various data sets
to produce the same output (assertions don't change).

This commit adds support for that, by checking if a test function
has input terminals. If that is the case, there must be a JSON
file next to the VI with the same file name, which defines one or
more data sets.

Example:

[
	{
		"name": "name of the data set",
		"data": {
			"control 1": "value",
			"control 2": "value",
			...
			"control n": "value"
		}
	}
]