Refactor Fact & Theory to use VI Refnum
Both Fact and Theory are currently build with a VI Path,
which results in repetitive creation of VI References to
validate the kind of VI and therefore kind of Test Case
(Fact or Theory).

To simplify the implementation, both Fact and Theory
should accept a VI Refnum instead of a VI Path.

This commit changes the implementation of Fact and Theory
to accept a VI Refnum as input and simplifies other helper
functions.