Add an error library
Constraints and some of the test VIs currently throw a custom error
with code 5000 to report any assertion failures.

While this approach works as intended, it requires everyone to use
the same error code for all failures. This is quite difficult to
maintain because there are no restrictions (it's an undocumented
agreement right now).

To prevent issues in the future, this library needs a standard way
to report errors - in particular assertion failures.

For this reason a new Error library is necessary, which provides
functions that can be used to report different kinds of errors. At
this time, the library has only one function to report expectation
failures (constraint failed their expectation):

- ExpectationFailedError

The error created by this VI uses the standard (undocumented)
LabVIEW error string to build a custom error message. The reason
this is not implemented with the standard 'Create Error Cluster
From Error Code' is because that function introduces sideeffects
and doesn't provide the necessary flexibility (i.e. the ability
to specify the caller manually).