Harmonize test standard error behavior
There are currently multiple tests that essentially do the
same operation for different assertions.

This can be harmonized into a single test that iterates over
all assertions.

This commit adds a new test 'testStandardErrorBehavior' that
goes over all assertions and checks if 'error out' returns
the same error that was provided to 'error in (no error)'.

Note that this also enforces the requirement for the error
terminals with default names (as provided by LabVIEW).

Perhaps there is a better way to do this in the future, so
that the output will have a single test for every assertion
instead of one test for all.