Run tests in separate instance
Create a new application instance in CodeCoverage and
provide the refnum to TestSuite in order to run tests
more efficiently.

Using a separate application instance is perhaps best
performance wise, because it is not affected by the
current instance and its settings.

Note that it is possible to access the application
instance from the code coverage module, which is not
a good solution, but the alternative would be to have
the code coverage module create test suites, which
doesn't make sense to me.