Fix incomplete coverage in test suite
With the recent change in the code coverage module,
the test suite was reported with a coverage percentage
of 90% for AddTestsFromProjectLibrary.vi (private).

As it turns out, code coverage for that method  was
higher previously because the same function is called
by the test runner.

This commit adds a fake theory to the fake library with
which the missing code section is covered.