Reduce ignore list
The test runner currently ignores itself and all test functions for
code coverage. This is no longer necessary because code coverage
now runs in a separate application instance.

This commit removes the test runner and all test libraries from the
ignore list and fixes any errors that occurred because of that.