Use a stream to speed up test runner
This commit adds a stream for the test runner to build test
suites while tests are running. This has the advantage that
loading files (which is quite slow) can be done while tests
are running in parallel. On the flip-side, test performance
is slowed slightly because loading and running VIs can lock
each other.

On average, the test should perform slightly faster than
before.