Change execution system for tests
The standard behavior for all tests is to run in the same execution
system as the caller. This can reduce performance, because the caller
uses Run VI to run the VI, which runs in the UI thread.

Tests should instead always run in the standard execution system to
get best performance (slight improvement). For the future maybe this
can be automated?