Add TestListener
TestListener is an interface to report the test status while the test
is running.

The TestListener currently supports two methods:

AddFailure - Reports that a test failed.
EndTest - Reports that a test ended.

Eventually, there will be a test explorer that needs to be informed at
runtime to present the data to the user. For this purpose, there is now
a TestExplorer library, that aims to capture all necessary libraries to
implement the user interface later. Note that these are not covered by
any unit tests right now!