Add ResultTree
ResultTree is a class that manages a tree control to display test results
including duration and error message. It is meant to be used eventually
by the test explorer (which doesn't exist yet).

ResultTree currently supports adding test suites (including nested suites)
and the ability to specify the test duration and error message.