Return parent tag name from ResultTree
ResultTree currently only supports adding elements to the tree, each of
which is identified by a tag. Nested items are represented by nested tags,
but there is no way to get the parent tag of a given tag.

This commit adds 'GetParent', which returns the parent tag of a given tag.