Properly manage nested libraries
TestSuite currently adds all tests from test libraries into a single suite
that includes all nested libraries. This has the disadvantage, that nested
libraries can not be used to organize tests for readability. Everything is
put under the umbrella of the top-most library. Additionally, changing the
access scope of a sub library to private has no effect.

To address these issues, TestSuite needs to traverse individual libraries
layer-by-layer and add them as nested suites. That will ensure, that every
library is managed as its own suite (which also makes it appear as nested
suite in the result tree) with the ability to organize tests in a more
readable form.

With this change, nested libraries are only added if they are in public
scope, therefore making it possible to 'disable' entire libraries (as was
already possible for individual tests).

This has some impact on the test time, because building nested test suites
requires more time. However, the benefit of having a more readable result
tree and the ability to manage tests are tremendous.