Replace EventTestListener by ResultTreeListener
The current way of updating the result tree is very inefficient. All the
result data needs to pass through custom user events, which are then
handled by a separate event loop to update the result tree.

A better way to do this is to have a specific listener for the result
tree and simply update the tree in-place.

This also removes the need for a separate event loop and the
EventTestListener altogether.