Correct list of not fully covered VIs
VIs that are never executed during test don't count towards the list
of not fully covered VIs. This is because the list is compile from
the output of the code coverage module, which is not a good solution.

A better solution is to create a complete list of VIs from a project
and remove all completely covered (100%) VIs from it. This will leave
VIs that are partially or not covered.