Remove error output from the UI
If there is an error, it is probably a good idea to inform the user about it
instead of silently displaying it on the front panel. This means that the
error indicator is uncessary. Simply display an error message.

With the error indicator gone, align the front panel controls, so that the
UI focuses on the result tree.