Set failure symbol to subtree
When a test fails, currently only the failed test gets the 'failed'
symbol. This makes it difficult to determine if a test failed, when
the project has lots of tests that don't fit easil in the result
tree.

A better approach is to not just fail the test, but also the owning
suite up to the root node. That way users can traverse test results
from the root elements and quickly determine which library contains
failures and which doesn't.