Handle failures in nested subtrees
There is currently a bug, where a failure in a subtree with failed
tests can prevent a sibling subtree from updating the 'passed' icon
because the flag 'Subtree Contains Failure' leaks into branches.

This can be solved by remembering the tag of the test that failed
last and then look if the current suite is a parent of that test.

Of course the tag can change between sibling branches, but it will
still work. Here are a few examples:

x parent
  x child_A
    x Test_A
    O Test_B
  O child_B
    O Test_A
    O Test_B
  X child_C
    X Test_A

In this case, the tag would change between

- parent/child_A/Test_A
- parent/child_A
- parent/child_C/Test_A
- parent/child_C
- parent
- <empty>

Only the libraries that actually contain a test failure will be marked
as failed.