Whitelist: ignore members of private libraries
Adding a library with private member libraries correctly ignores the
private library, but not functions within private libraries.

This is a limitation of GetAllDescendants, which traverses into all
sublibraries. The solution is to use GetAllDescendants2 to exclude
sub libraries, and traverse through each sub library manually.

This increases execution time for projects with few private libraries
and speeds up execution for projects with many files inside private
libraries. In any case, the fixed behavior is correct.