Fix Test Explorer shows items from last save on launch
On every launch of Test Explorer, it displays the values of when the
VI was last saved. This is a known behavior of LabVIEW, which can be
solve in two ways:

1. Clear the tree at the end of execution
2. Clear the tree at start of execution

Implemented 1. to guarantee correct behavior in all cases. The other
option does not seem to work reliably (values are still visible)