Show active project in Test Explorer
When working with multiple projects or running Test Explorer without
any project, it is not always clear to the user which project (if any)
is affected.

This commit adds an indicator to Test Explorer, which displays the
currently active project. If the user opens or selects a different
project, the indicator updates automatically.