Fix error 5000 occurred at BuildTestFromProjectItem.vi
This error occurs when LabVIEW Unit Test Framework Toolkit is not
installed, because '.lvtest' and '.lvvect' files are reported with
type string '{00000000-0000-0000-0000-000000000000}' instead of
'TestItem' and 'TestVectorItem'.

Fixed by adding '{00000000-0000-0000-0000-000000000000}' to the list
of ignored items. This should also prevent other types of files from
being flagged when the toolkit is not installed.

Fixes #4