Rename TestExplorer to Test Explorer (LVUnit)
The tools menu item "TestExplorer" does not provide much context for
the user and does not use normal text. Changing the name to
"Test Explorer (LVUnit)" will make it easier to understand.

Note: I have decided against a subcategory because it is only one item
and a subcategory only hides it unnecessarily. This might be an option
for the future.

Closes #10