Add quick-drop plugin for Test Explorer
Use Ctrl-E to launch Test Explorer.

Important note: in order to assign a different shortcut, VIPM needs
to provide a way to disable changing VI documentation for some VIs.
Currently it also changes the description of the QD plugin, which
results in the shortcut being set to the second-last character in the
description "e" => "E".

Closes #17