Discover tests prior to execution
Use the project refnum to discover tests and display the total number
of tests when a project is selected in the test explorer.

Important: discovery of theories is not implemented, which means that
the total number of tests displayed does not add up with the number of
failed and passed tests after execution.