Prepare and execute tests in parallel
Test preparation can happen in a parallelized for-loop because it has
no impact on the order of execution. Test execution itself, while done
sequentially, can start as soon as the first test is prepared.

Depending on the number of threads available, performance is improved
by a lot (tm).

Important: if tests could run in parallel, performance should improve
by approx. 1/3 on average. Perhaps this can become an option in the
future. Note that tests are not run in parallel by default because it
could cause side-effects and pseudo-failures, in particular when doing
integration tests.