Remove control data from test case
For parameterized tests, the control data must be constructed based
on the data types of each control. The data types, however, are not
known at the time a test case is constructed. Thus, the control data
is not defined by the base class for test cases.

For specific unit test frameworks, the test case can be subclassed
to provide control data in the framework-specific format.