Restore functionality to pass all tests
1) Subclass Test Case for LVUnit

This adds a subclass for test cases managed by LVUnit, which allows
it to pass JSON data as part of a test case. The JSON data contains
the control data in JSON format.

LVUnit Test Executor uses the JSON data to prepare the VI for execution.

2) Use JSONtext Object Serialization for class data

Objects are currently reported in binary format, which makes it very
difficult to know what the difference is.

This adds a custom serializer to allow serializing and parsing of
objects directly from JSON.