Test Result: Add start and end time
This adds Start Time and End Time to Test Result. These times are
used to calculate the duration for every test case.

The duration shown in Test Explorer is correct for individual test
cases, but always displays the same duration as the last test case
in the column for the parent test suite.