Constraint: Fully decouple test framework from assertion library
This introduces a 'Test Framework Factory' that decouples the specific
test framework implementation from the assertion library.

The active test framework is stored in a global configuration that can
also be specified directly by the caller. If no test framework is
specified, the framework is detected from the application domain.

Currently, the detection in the application domain is not functional and
always returns the LVUnit Test Framework.

All supported test frameworks are defined in the enumeration 'Test
Framework.ctl'. "Unknown" defaults to the application domain.