Framework: Add Caraya Test Framework
This adds Caraya as a test framework option. It is used when 'Define
Test' is found in the application domain.

Unfortunately, it does not seem possible to completely decouple the
Framework library from Caraya because Caraya uses Call Chain to
register assertions and Call By Reference always runs as top-level VI.

This means that LVUnit (for now) has a hard dependency on Caraya. I hope
it will be possible to decouple the two in the future.