Framework: Completely decouple Caraya
This completely decouples Caraya from the assertion library by
dynamically loading the Caraya Test Framework class on-demand when
Caraya is detected in the application context.

It is very important that the class is not added to the project or
otherwise included in a library. Perhaps we can find a solution that
makes it easier to understand?

In any case, Caraya is correctly loaded and executed without any issues
regarding Call Chain. Thus, it is completely decoupled. This now also
opens up the possibility to use other unit test frameworks :)