Execute tests in the application instance of the project
When Test Explorer is started from the Tools Menu, it runs in a special
application context for menu items. This makes it impossible to capture
performance statistics on VIs.

This changes the test executor to load all VIs in the application
context of the owning project.

As a side-effect, this also improves execution speed because the VIs
under test do not need to be loaded into a new context.