Add 'Performance Mode'
This adds a new 'Performance Mode' option to Test Explorer, which
determines whether UI updates occur live or only at the end of a test.

By default, the option is turned on. This is the standard behavior that
was previously used. When turned off, the UI is updated live and test
execution will take considerably longer.