Add code coverage support (Cobertura)
This adds support for exporting code coverage in Cobertura format.

Code coverage is determined from the application context and converted
into Cobertura format. Branch rate is the number of diagrams divided by
the number of uncovered diagrams. Line rate is the percentage value
reported by the code coverage engine.

Packages are determined from the fully-qualified name of a VI. A class
in Cobertura represents a VI in LabVIEW.