Quick Drop: Fix default shortcut changes from 'E' to 'S'
The default shortcut for the Quick Drop action 'Start Test Explorer' is
supposed to default to 'E'. However, when it is installed with VIPM, the
default shortcut changes to 'S'.

The reason for this is the SPDX-License-Identifier that is added during
the build step of the package. The default shortcut 'S' is because of
the 'S' of 'SPDX'.

For the default shortcut to work as intended, the line with "Default
Shortcut - [E]" must be at the end of the VI description.

Since there is no way in VIPM to selectively prevent VI descriptions
from being changed, this can only be fixed in a post-install step.

This adds a post-install step to remove any text after the default
shortcut.