Fix Test Explorer opens VIs in the context of the Tools menu
When Test Explorer is launched from Tools Menu or Quick Drop, its 
application refnum points to the application context of the Tools Menu 
or Quick Drop, respectively. This results in VIs being opened multiple 
times for each different application context when selecting a VI from 
the tree.

This fixes the behavior by opening the VI in the application context of 
the active project.